more renaming and moving
