+++ packages/back-end/src/poke-api/index.ts
@@ -1,0 +1,3 @@
+export { NamedAPIResource } from './named-api-resource'
+export { NamedAPIResourceList } from './named-api-resource-list'
+export { PokeApiPokemon } from './poke-api-pokemon'
